// marketplace-fe/src/app/features/market-factory/enums/market-enums.ts
export enum MarketEvents {
    MARKET_CREATED = 'market.created',
    MARKET_SAVED = 'market.saved',
    MARKET_UPDATED = 'market.updated',
    MARKET_DELETED = 'market.deleted',
    MARKET_ACCEPTED = 'market.accepted',
    MARKET_REJECTED = 'market.rejected',
    MARKET_ACTIVATED = 'market.activated',
    MARKET_DEACTIVATED = 'market.deactivated',
    MARKET_SUBSCRIBED = 'market.subscribed'
}

export enum MarketState {
    CREATED_OFFLINE = 'CREATED_OFFLINE',
    CREATED_OFFLINE_PENDING_ACCEPTATION = 'CREATED_OFFLINE_PENDING_ACCEPTATION',
    CREATED_OFFLINE_ACCEPTED = 'CREATED_OFFLINE_ACCEPTED',
    CREATED_OFFLINE_REJECTED = 'CREATED_OFFLINE_REJECTED',
    CREATED_OFFLINE_EXPIRED = 'CREATED_OFFLINE_EXPIRED',
    WALLET_CREATED_PENDING_ACTIVATION = 'WALLET_CREATED_PENDING_ACTIVATION',
    ACTIVE_ONCHAIN = 'ACTIVE_ONCHAIN',
    SUSPENDED = 'SUSPENDED',
    DEACTIVATED = 'DEACTIVATED'
}